pyproject.toml
[project]
name = "api-testing-demo"
version = "0.1.0"
description = "Demo of REST API testing"
requires-python = ">=3.11"
dependencies = [
    "pre-commit==4.2.0",
    "pytest>=8.3.5",
    "pytest-html>=4.1.1",
    "requests>=2.32.3",
    "pydantic>=2.0.0",
    "pytest-dotenv==0.5.2",
]
